src/requirements.txt
customtkinter
tk
pillow
CTkSpinbox
ctkcomponents
markdown
tkinterweb
